requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
crispy-bootstrap5==2024.2
Django==4.2.13
django-bootstrap-v5==1.0.11
django-crispy-forms==2.1
django-redis==5.4.0
numpy==1.26.4
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.4
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
whitenoise==6.6.0
